// src/types/acompanhamento.ts
export type Acompanhamento = {
  codigo?: number
  data?: string
  hora?: string
  peso?: string
  altura?: string
  pasistolica?: string
  padiastolica?: string
  freqresp?: string
  freqcard?: string
  tempcorporal?: string
  hgt?: string
  saturacao?: string
  circabdominal?: string
  perimetrocefalico?: string
  statusvacina?: string
  observacoes?: string
  datastr?: string
  nomeprofissional?: string
}